Refactor uploads
